refactor: utility가 적은 함수 co-locating
